Move test from test_playwright.py to test_a11y.py
